DeepLearning-chatbot/requirements.txt
fastapi
uvicorn
pandas
sentence-transformers
google-generativeai
rank_bm25
jinja2
